Automatically generate the bootstrap config documentation
The bootstrap config documentation may get out of sync as a result
of apiv1.BootstrapConfig changes.

To avoid this, we're going to automatically generate the documentation
based on the apiv1.BootstrapConfig structure definition and docstrings.

This change adds the necessary tooling as part of the ``docgen``
package, which may be reused for other structures or projects.